Deixando os lembretes realmente recoorentes

+++ src/main/java/br/com/eric/telegram/kerobot/scheduler/Scheduler.java
@@ -23,7 +23,9 @@ public class Scheduler {
 		Iterable<br.com.eric.telegram.kerobot.models.Scheduled> scheduleds = scheduledRepository.findAllByTimeLessThan(new Date().getTime());
 		scheduleds.forEach(s -> {
 			reminderExecutor.execute(s);
-			scheduledRepository.delete(s);
+			if (!s.isFrequently()) {
+				scheduledRepository.delete(s);
+			}
 		});
 	}
